Update number_of_islands_leetcode_200.py
In  LeetCode problem "Number of Islands", the problem asks you to create an algorithm that counts the number of islands in a given 2D binary grid representing a map of '1's (land) and '0's (water). Islands are defined as connected regions of '1's, where connections can be horizontal or vertical, and are surrounded by water on all sides.
In my solution, I implemented a breadth-first search (BFS) algorithm to traverse the neighboring nodes of each of the nodes with value == '1' and identify the boundaries of the corresponding island. For each '1' node encountered during the iteration of the 2D list argument "grid", the node's position [i, j] is passed to a helper function called "getIslandLength".
The BFS is implemented using a stack named “stack”. While the stack is not empty, the function pops an element and checks its neighbors. If a neighbor is in the bounds of the grid and is also of value '1', it is appended to the stack for processing. To avoid revisiting the same node, we mark visited nodes with a special character '*'.
Once the BFS is complete for the node passed in to getIslandLength function, we return to the main loop and increment the "numIslands" counter. This variable keeps track of the total number of islands identified in the grid.

After the iteration of “grid” is complete I return the numIslands variable which represents the total number of islands found.